Moves app/common/ components folder to global/components/ and updates paths accordingly
