+++ src/app/about/recommendations/Recommendations.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { WebLink } from '../../common/WebLink'
+import { WebLink } from '../../../global/components/WebLink'
 
 export const Recommendations = () => {
   return (
