+++ src/app/portfolio/featured/FeaturedContentCard.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { WebLink } from '../../common/WebLink'
+import { WebLink } from '../../../global/components/WebLink'
 
 export const FeaturedContentCard = ({ featuredContentItem }) => {
   const card = (
